+++ test/snek/board/size_test.exs
@@ -1,0 +1,5 @@
+defmodule SizeTest do
+  use ExUnit.Case
+  alias Snek.Board.Size
+  doctest Size, import: true
+end