+++ test/snek/board/snake_test.exs
@@ -1,0 +1,5 @@
+defmodule SnakeTest do
+  use ExUnit.Case
+  alias Snek.Board.Snake
+  doctest Snake, import: true
+end